move some client only classes to client package
