+++ src/main/java/com/github/thedeathlycow/frostiful/client/registry/FParticleFactoryRegistry.java
@@ -1,5 +1,8 @@
-package com.github.thedeathlycow.frostiful.particle.client;
+package com.github.thedeathlycow.frostiful.client.registry;
 
+import com.github.thedeathlycow.frostiful.Frostiful;
+import com.github.thedeathlycow.frostiful.client.particle.HeatDrainParticle;
+import com.github.thedeathlycow.frostiful.client.particle.WindParticle;
 import com.github.thedeathlycow.frostiful.registry.FParticleTypes;
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
@@ -10,7 +13,8 @@ import net.minecraft.particle.ParticleType;
 @Environment(EnvType.CLIENT)
 public class FParticleFactoryRegistry {
 
-    public static void registerFactories() {
+    public static void initialize() {
+        Frostiful.LOGGER.debug("Initialized Frostiful particle factories");
         registerFactory(FParticleTypes.HEAT_DRAIN, HeatDrainParticle.Factory::new);
         registerFactory(FParticleTypes.WIND, WindParticle.Factory::new);
         registerFactory(FParticleTypes.WIND_FLIPPED, WindParticle.Factory::new);
