+++ src/main/java/com/github/thedeathlycow/frostiful/mixins/client/SpriteAtlasManagerMixin.java
@@ -1,5 +1,6 @@
 package com.github.thedeathlycow.frostiful.mixins.client;
 
+import com.github.thedeathlycow.frostiful.Frostiful;
 import com.github.thedeathlycow.frostiful.client.registry.FTexturedRenderLayers;
 import net.minecraft.client.render.model.SpriteAtlasManager;
 import net.minecraft.util.Identifier;
@@ -19,6 +20,7 @@ public class SpriteAtlasManagerMixin {
             at = @At("HEAD")
     )
     private static Map<Identifier, Identifier> setAtlases(Map<Identifier, Identifier> loaders) {
+        Frostiful.LOGGER.debug("Registered Frostiful atlas loaders");
         Map<Identifier, Identifier> mutableLoaders = new HashMap<>();
         mutableLoaders.putAll(FTexturedRenderLayers.LAYERS_TO_LOADERS);
         mutableLoaders.putAll(loaders);
